requirements.txt
flask
flask-cors
python-telegram-bot==20.7
gunicorn
